requirements.txt
absl-py==0.14.0
affine==2.3.0
astunparse==1.6.3
attrs==21.2.0
Automat==0.8.0
awscli==1.22.5
blinker==1.4
boto3==1.20.5
botocore==1.23.5
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.15.2
ec2-hibinit-agent==1.0.0
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
GDAL==3.3.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
hibagent==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
incremental==16.10.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.8.0
keras-gradient-noise==0.11
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.4.3
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.6.3
numpy==1.22.2
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
pandas==1.3.3
pexpect==4.6.0
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==4.0.32
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
rasterio==1.2.8
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.7.2
s3transfer==0.5.0
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
snuggs==1.4.7
sos==4.3
ssh-import-id==5.10
systemd-python==234
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-estimator==2.6.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==2.2.0
tifffile==2021.8.30
Twisted==18.9.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.7
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
wadllib==1.3.3
Werkzeug==2.0.1
whitebox==2.0.3
wrapt==1.12.1
zipp==3.5.0
zope.interface==4.7.1
